Remove use of `CPUSummary`
The use of CPUInfo makes `--trim` difficult.
Removing this dependency here would unlock a large amount of libraries
which use the Polyester library to be trimmmable (notably e.g. almost
everything in the SciML ecosystem).

However, we might need a bit more discussion on the exact removal of
this feature.